# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mobile App API",
    "description": "API documentation for the mobile app, including orders, products, service centers, reviews, and more.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "components": {
    "schemas": {
...
      "Order": {
        "type": "object",
        "required": [
          "user",
          "products",
          "totalAmount",
          "paymentStatus",
          "orderStatus",
          "deliveryAddress",
          "deliveryDate"
        ],
        "properties": {
          "user": {
            "type": "string",
            "description": "User ID"
          },
          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string",
                  "description": "Product ID"
                },
                "quantity": {
                  "type": "number",
                  "description": "Quantity of the product"
                },
                "price": {
                  "type": "number",
                  "description": "Price of the product"
                }
              }
            }
          },
          "totalAmount": {
            "type": "number",
            "description": "Total amount of the order"
          },
          "paymentStatus": {
            "type": "string",
            "enum": ["pending", "completed", "failed"],
            "description": "Payment status"
          },
          "orderStatus": {
            "type": "string",
            "enum": ["processing", "shipped", "delivered"],
            "description": "Order status"
          },
          "deliveryAddress": {
            "type": "string",
            "description": "Delivery address"
          },
          "deliveryDate": {
            "type": "string",
            "format": "date",
            "description": "Delivery date"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/orders": {
      "post": {
        "summary": "Place a new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order placed successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order details by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/orders/user/{userId}": {
      "get": {
        "summary": "Get all orders for a specific user",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders for the user"
          },
          "404": {
            "description": "No orders found for this user"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}